Implement main.py  structure with web engine and score system
