+++ web/index.html
@@ -94,10 +94,12 @@
   <div id="t-ico">🏆</div>
   <div><div id="t-name"></div><div id="t-desc"></div></div>
 </div>
-
-<script src="js/phaser.min.js"></script>
-<script type="text/javascript" src="js/script.js"></script>
 <script src="js/qwebchannel.js"></script> 
+<script src="js/phaser.min.js"></script>
+<script type="text/javascript" src="js/score-bridge.js"></script>
+<script type="text/javascript" src="js/script.js"></script>
+<script type="text/javascript" src="js/scene.js"></script>
+
 	
 </body>
 </html>
